finished responsive header
