+++ src/app/components/Header/style.module.css
@@ -1,4 +1,4 @@
-.header, .nav-menu, .icons-wrapper {
+.header {
     display: flex;
     justify-content: space-between;
     align-items: center;
@@ -10,26 +10,10 @@
     border: 1px solid var(--olive);
     padding: 0.2% 6%;
 }
-.nav-menu {
-    display: flex;
-    justify-content: space-between;
-}
-.menu-item {
-    list-style: none;
-    font-size: 1.4rem;
-    font-weight: 500;
-}
-.menu-item a {
-    text-decoration: none;
-    color: var(--darkGrey);
-}
-.icons-wrapper {
-    gap: 1rem;
-}
-.icon {
-    cursor: pointer;
-}
-.icon svg {
-    width: 1.2rem;
-    height: 1.2rem;
-}
+
+/* @media (max-width: 1200px) {
+   .header {
+        padding: 0 10%;
+   }
+} */
+
