+++ src/app/components/Logo/style.module.css
@@ -1,0 +1,29 @@
+
+.logo svg {
+    width: 10rem;
+}
+
+@media (max-width: 600px) {
+    .logo svg {
+        width: 5rem;
+    }
+    .logo {
+        margin-left: 10%;
+    }
+
+}
+@media (min-width:600px) and (max-width: 900px) {
+    .logo svg {
+        width: 8rem;
+    }
+    .logo {
+        margin-left: 10%;
+    }
+
+}
+@media (min-width:900px) and (max-width: 1200px) {
+    .logo {
+        margin-left: 10%;
+    }
+
+}
